refactor(standard_layout): Change chicken balance to float

+++ internal/domain/farmer_repository.go
@@ -3,7 +3,7 @@ package domain
 import "github.com/google/uuid"
 
 type FarmerRepository interface {
-	CreateFarmer(name string, chickenBalance int32, cashBalance int32) (*Farmer, error)
+	CreateFarmer(name string, chickenBalance float64, cashBalance int32) (*Farmer, error)
 	GetFarmerByName(string) (*Farmer, error)
 	GetFarmers() ([]Farmer, error)
 	DeleteFarmerByID(ID uuid.UUID) error
